app/backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
pymongo
websockets
requests
beautifulsoup4
lxml
groq